mobile view for event signups

+++ src/components/Admin/Events/EventSignupTable.tsx
@@ -1,11 +1,12 @@
 import React, {Component} from 'react'
-import { Table, Tooltip, notification, Button } from 'antd'
-import { FormOutlined, ArrowRightOutlined, ArrowLeftOutlined } from '@ant-design/icons'
+import { Table, Tooltip, notification, Button, Select, Tag } from 'antd'
+import { FormOutlined, CopyOutlined, ArrowRightOutlined, ArrowLeftOutlined } from '@ant-design/icons'
 import './EventSignupTable.css'
 import { AumtEventSignupData, AumtEventSignup } from '../../../types'
 import moment from 'moment'
 import { ColumnsType } from 'antd/lib/table'
 import db from '../../../services/db'
+import dataUtil from '../../../services/data.util'
 
 
 type TableRow = AumtEventSignupData & {key: string, displayTime: string}
@@ -20,6 +21,7 @@ interface EventSignupTableState {
     rows: TableRow[]
     columns: ColumnsType<TableRow>
     tableLoading: boolean
+    selectedSignup: TableRow | null
 }
 
 export class EventSignupTable extends Component<EventSignupTableProps, EventSignupTableState> {
@@ -28,7 +30,8 @@ export class EventSignupTable extends Component<EventSignupTableProps, EventSign
         this.state = {
             rows: [],
             columns: [],
-            tableLoading: false
+            tableLoading: false,
+            selectedSignup: null
         }
     }
     componentDidMount = () => {
@@ -45,7 +48,10 @@ export class EventSignupTable extends Component<EventSignupTableProps, EventSign
         }
     }
 
-    updateConfirmed = (row: TableRow, newConfirmed: boolean) => {
+    updateConfirmed = (row: TableRow | null, newConfirmed: boolean) => {
+        if (!row) {
+            return
+        }
         db.confirmMemberEventSignup(this.props.eventId, row.key, newConfirmed, this.props.isWaitlist)
             .then(() => {
                 notification.success({message: `Updated confirmed for ${row.displayName} to ${newConfirmed ? 'Yes' : 'No'}`})
@@ -55,6 +61,7 @@ export class EventSignupTable extends Component<EventSignupTableProps, EventSign
             })
     }
     deleteMember = (key: string) => {
+        if (!key) return
         db.removeMemberFromEvent(key, this.props.eventId, this.props.isWaitlist)
             .then(() => {
                 notification.success({message: `Removed from ${this.props.isWaitlist ? 'waitlist' : 'signups'}`})
@@ -64,6 +71,7 @@ export class EventSignupTable extends Component<EventSignupTableProps, EventSign
             })
     }
     onMoveClick = (key: string) => {
+        if (!key) return
         this.setState({...this.state, tableLoading: true})
         const data = this.props.signupData[key]
         db.signUpToEvent(this.props.eventId, key, data, !this.props.isWaitlist)
@@ -80,6 +88,15 @@ export class EventSignupTable extends Component<EventSignupTableProps, EventSign
                 this.setState({...this.state, tableLoading: false})
             })
     }
+    onSelectSignup = (key: string) => {
+        const member = this.state.rows.find(r => r.key === key)
+        if (member) {
+            this.setState({
+                ...this.state,
+                selectedSignup: member
+            })
+        }
+    }
     handleNewData = (signupData: AumtEventSignup) => {
         const rows = Object.keys(signupData)
             .map((uid: string) => {
@@ -90,10 +107,12 @@ export class EventSignupTable extends Component<EventSignupTableProps, EventSign
                         displayTime: moment(new Date(signup.timeSignedUpMs)).format('MMM DD HH:mm')
                     })
             })
+        const selectedMember = rows.find(r => r.key === this.state.selectedSignup?.key)
         this.setState({
             ...this.state,
             rows,
-            columns: this.getColumns()
+            columns: this.getColumns(),
+            selectedSignup: selectedMember || null
         })
     }
     getColumns = () => {
@@ -147,7 +166,51 @@ export class EventSignupTable extends Component<EventSignupTableProps, EventSign
         ]
         return columns
     }
+    copyText = (text: string | undefined) => {
+        if (text) dataUtil.copyText(text)
+    }
+    onMemberSelect = (member: TableRow) => {
+        this.setState({...this.state, selectedSignup: member})
+    }
     render() {
+        if (window.innerWidth < 600) {
+            return <div>
+                <Select
+                className='eventSignupSelect'
+                placeholder='Select Member'
+                onChange={this.onSelectSignup}
+                >
+                    {this.state.rows.sort((a, b) => a.timeSignedUpMs - b.timeSignedUpMs).map((signup: TableRow) => {
+                            return <Select.Option key={signup.key} value={signup.key}>
+                                <div className='eventSignupSelectName'>{signup.displayName}</div>
+                                <div className='eventSignupSelectTime'>{signup.displayTime}</div>
+                                {signup.confirmed ? 
+                                    <Tag className='eventSignupSelectPaidTag' color='success'>Paid</Tag>
+                                    : ''}
+                                </Select.Option>
+                        })}
+                </Select>
+                {this.state.selectedSignup ?
+                <div className="eventSignupMemberInfo">
+                    <p>Email: {this.state.selectedSignup.email} <CopyOutlined onClick={e => this.copyText(this.state.selectedSignup?.email)}/></p>
+                    <p>Paid: {this.state.selectedSignup.confirmed ? 'Yes' : 'No'}
+                        <Button 
+                        type='link'
+                        onClick={e => this.updateConfirmed(this.state.selectedSignup, !this.state.selectedSignup?.confirmed)}>
+                            Change
+                            </Button>
+                        </p>
+                    <Button
+                        onClick={e => this.onMoveClick(this.state.selectedSignup?.key || '')}
+                        disabled={!this.state.selectedSignup}>
+                        Move to {this.props.isWaitlist ? ' signups' : ' waitlist'}</Button>
+                    <Button
+                        onClick={e => this.deleteMember(this.state.selectedSignup?.key || '')}
+                        disabled={!this.state.selectedSignup} type='link'>Delete</Button>
+                </div>
+                : ''}
+            </div>
+        }
         return <div>
             <Table
                 size='small'
